Symlink `shared` packages
