+++ operational/core/src/test/kotlin/io/github/oxidefrp/core/shared
@@ -1,0 +1,1 @@
+../../../../../../../../../denotational/core/src/test/kotlin/io/github/oxidefrp/core/shared
