--- operational/core/src/test/kotlin/io/github/oxidefrp/core/test_framework/shared/utils.kt
@@ -1,4 +1,0 @@
-package io.github.oxidefrp.core.test_framework.shared
-
-inline fun <E, R : Comparable<R>> List<E>.isMonotonicallyIncreasingBy(crossinline selector: (E) -> R): Boolean =
-    zipWithNext { a, b -> selector(a) < selector(b) }.none { isIncreasing -> !isIncreasing }
